move research pages to research folder and clean out PRIVATE_SENSITIVE_DATA stuff
