+++ client/src/research/Analyses.js
@@ -1,9 +1,9 @@
 import Filters from './Filters.js';
 import _ from 'lodash';
-import LATEST_CONSENT from './PRIVATE_SENSITIVE_DATA/consent-teacher-moments/consented-latest.json';
-import HMTCA_PAPER_CONSENT from './PRIVATE_SENSITIVE_DATA/consent-teacher-moments/consented-hmtca-paper.json';
-const LATEST_DB_PATH = '/PRIVATE_SENSITIVE_DATA/db/db-latest.json';
-const LATEST_S3_PATH = '/PRIVATE_SENSITIVE_DATA/s3/2018-02-12/';
+import LATEST_CONSENT from './consented-latest.json';
+import HMTCA_PAPER_CONSENT from './consented-hmtca-paper.json';
+const LATEST_DB_PATH = '';
+const LATEST_S3_PATH = '';
 
 
 function isForUnconsciousBias(row) {
