+++ client/src/research/EmailLinkLoginPage.js
